// src/app/interfaces/categoria-evento.enum.ts
export enum CategoriaEvento {
    ANIVERSARIO = 'ANIVERSARIO',
    CITA = 'CITA',
    COMPRAS = 'COMPRAS',
    CUIDADO_PERSONAL = 'CUIDADO_PERSONAL',
    CUMPLEANIOS = 'CUMPLEANIOS',
    ENTRETENIMIENTO = 'ENTRETENIMIENTO',
    ESTUDIO = 'ESTUDIO',
    HOGAR = 'HOGAR',
    LIMPIEZA = 'LIMPIEZA',
    MEDICO = 'MEDICO',
    OTRO = 'OTRO',
    PAGO = 'PAGO',
    REUNION = 'REUNION',
    SALUD = 'SALUD',
    SOCIAL = 'SOCIAL',
    TRABAJO = 'TRABAJO',
    VIAJE = 'VIAJE'
}